2009-02-12	Joel Sherrill <joel.sherrill@oarcorp.com>
	* cpu.c: Change prototype of IDLE thread to consistently return void *
	and take a uintptr_t argument.

+++ cpukit/score/cpu/c4x/cpu.c
@@ -163,7 +163,7 @@ void _CPU_ISR_install_vector(
  */
 
 #if (CPU_PROVIDES_IDLE_THREAD_BODY == 1)
-void *_CPU_Thread_Idle_body( uint32_t ignored )
+void *_CPU_Thread_Idle_body( uintptr_t ignored )
 {
 
   for( ; ; ) {
