+++ cpukit/score/cpu/h8300/cpu.c
@@ -143,7 +143,7 @@ void _CPU_Install_interrupt_stack( void )
  */
 
 #if 0
-void *_CPU_Thread_Idle_body( uint32_t ignored )
+void *_CPU_Thread_Idle_body( uintptr_t ignored )
 {
 
   for( ; ; )
